auxdisplay: Extract character line display core support
Extract the character line display core support from the simple ASCII
LCD driver for the MIPS Boston, Malta & SEAD3 development boards into
its own subdriver, so it can be reused for other displays.

Note that this moves the "message" device attribute in sysfs in a
"linedisp.N" subdirectory.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>
---
Changes to img-ascii-lcd.c are untested due to lack of hardware.